requirements.txt
# Required pip install

python-telegram-bot
pyttsx3
